refactor: optimize linkedin topic recommendation service
