+++ backend/services/integrations/linkedin/topic_recommendation_service.py
@@ -2,11 +2,17 @@
 Phase 6 — topic recommendation orchestration (cache-first generate).
 
 Gates on profile completeness and AI intelligence presence; coordinates LLM + validation + persistence.
+
+Uses the centralized llm_text_gen service for all LLM calls, ensuring:
+- Subscription checking and usage tracking
+- Provider routing and fallback handling
+- Routing observability
 """
 
 from __future__ import annotations
 
-from typing import Any, Literal, Optional, TypedDict
+import json
+from typing import Any, Callable, Literal, Optional, TypedDict
 
 from loguru import logger
 
@@ -19,10 +25,8 @@ from services.integrations.linkedin.profile_repository import (
     compute_ai_intelligence_hash,
 )
 from services.integrations.linkedin.profile_validation_types import ProfileValidationResult
-from services.integrations.linkedin.topic_recommendation_llm import (
-    TopicRecommendationGenerateFn,
-    TopicRecommendationLLMError,
-    call_topic_recommendation_llm,
+from services.integrations.linkedin.topic_recommendation_types import (
+    topic_recommendation_json_schema,
 )
 from services.integrations.linkedin.topic_recommendation_validator import (
     TopicRecommendationValidationError,
@@ -31,7 +35,7 @@ from services.integrations.linkedin.topic_recommendation_validator import (
     normalize_topic_recommendation_raw,
     validate_topic_recommendation_payload,
 )
-from services.llm_providers.gemini_provider import gemini_structured_json_response
+from services.llm_providers.main_text_generation import llm_text_gen
 
 _LOG_PREFIX = "[TopicRecommendation]"
 
@@ -40,9 +44,20 @@ VALIDATION_RETRY_USER_SUFFIX = (
     "Return valid JSON only matching the required schema exactly with exactly five recommendations."
 )
 
+# Type alias for the LLM generation function (for test injection)
+TopicRecommendationGenerateFn = Callable[..., str]
+
 
 class TopicRecommendationError(Exception):
     """Base error for topic recommendation orchestration."""
+
+
+class TopicRecommendationLLMError(Exception):
+    """Raised when the LLM provider fails for topic recommendation generation."""
+
+    def __init__(self, message: str, *, error_kind: str = "unknown") -> None:
+        super().__init__(message)
+        self.error_kind = error_kind
 
 
 class TopicRecommendationAcquireMeta(TypedDict, total=False):
@@ -59,7 +74,7 @@ def get_or_generate_topic_recommendations(
     profile_validation: ProfileValidationResult,
     repository: Optional[ProfileRepository] = None,
     force_regenerate: bool = False,
-    generate_fn: TopicRecommendationGenerateFn = gemini_structured_json_response,
+    generate_fn: Optional[TopicRecommendationGenerateFn] = None,
 ) -> tuple[Optional[list[dict[str, Any]]], TopicRecommendationAcquireMeta]:
     """
     Return cached or newly generated topic recommendations.
@@ -334,20 +349,117 @@ def _generate_and_persist_recommendations(
     return stored
 
 
+def _classify_llm_error(exc: Exception) -> str:
+    """Return a safe error category for LLM/provider failures (for logging)."""
+    msg = str(exc).lower()
+    if "resource_exhausted" in msg or "quota" in msg or "rate limit" in msg:
+        return "quota_or_rate_limit"
+    if "401" in msg or "403" in msg or "api key" in msg or "authentication" in msg:
+        return "auth"
+    if "timeout" in msg or "timed out" in msg or "deadline" in msg:
+        return "timeout"
+    if "invalid json" in msg or "json" in msg:
+        return "invalid_json"
+    if "subscription" in msg or "429" in msg:
+        return "subscription_limit"
+    return "provider_error"
+
+
 def _call_llm_with_validation_retry(
     *,
     user_id: str,
     user_prompt: str,
-    generate_fn: TopicRecommendationGenerateFn,
+    generate_fn: Optional[TopicRecommendationGenerateFn] = None,
 ) -> dict[str, Any]:
-    """Call LLM once; retry once when validation fails on the parsed response."""
+    """Call LLM once; retry once when validation fails on the parsed response.
+
+    Uses llm_text_gen for subscription checking, usage tracking, and provider routing.
+    """
     logger.info("{} Sending request to LLM user_id={}", _LOG_PREFIX, user_id)
-    raw = call_topic_recommendation_llm(
-        system_prompt=TOPIC_RECOMMENDATION_SYSTEM_PROMPT,
-        user_prompt=user_prompt,
-        user_id=user_id,
-        generate_fn=generate_fn,
-    )
+
+    schema = topic_recommendation_json_schema()
+
+    def _call_llm(prompt: str) -> dict[str, Any]:
+        """Inner helper to call LLM and parse JSON response."""
+        try:
+            # Use the centralized llm_text_gen which handles:
+            # - Subscription preflight checks
+            # - Provider routing and fallbacks
+            # - Usage tracking
+            # - Routing observability
+            if generate_fn is not None:
+                # Test path: use injected mock function
+                response_text = generate_fn(
+                    prompt=prompt,
+                    system_prompt=TOPIC_RECOMMENDATION_SYSTEM_PROMPT,
+                    json_struct=schema,
+                    user_id=user_id,
+                )
+            else:
+                # Production path: use centralized service
+                response_text = llm_text_gen(
+                    prompt=prompt,
+                    system_prompt=TOPIC_RECOMMENDATION_SYSTEM_PROMPT,
+                    json_struct=schema,
+                    user_id=user_id,
+                    flow_type="linkedin_topic_recommendations",
+                )
+
+            # Parse JSON response
+            if isinstance(response_text, dict):
+                return response_text
+            if isinstance(response_text, str):
+                try:
+                    parsed = json.loads(response_text)
+                    if not isinstance(parsed, dict):
+                        raise TopicRecommendationLLMError(
+                            "LLM JSON response must be an object",
+                            error_kind="invalid_json",
+                        )
+                    if parsed.get("error"):
+                        error_message = str(parsed.get("error"))
+                        logger.error(
+                            "{} LLM returned error payload message={}",
+                            _LOG_PREFIX,
+                            error_message[:500],
+                        )
+                        raise TopicRecommendationLLMError(
+                            "LLM provider returned an error response",
+                            error_kind="provider_error",
+                        )
+                    return parsed
+                except json.JSONDecodeError as exc:
+                    logger.error(
+                        "{} LLM returned invalid JSON string error={}",
+                        _LOG_PREFIX,
+                        exc,
+                    )
+                    raise TopicRecommendationLLMError(
+                        "LLM returned invalid JSON string",
+                        error_kind="invalid_json",
+                    ) from exc
+            raise TopicRecommendationLLMError(
+                f"Unexpected LLM response type: {type(response_text).__name__}",
+                error_kind="invalid_response",
+            )
+        except TopicRecommendationLLMError:
+            raise
+        except Exception as exc:
+            error_kind = _classify_llm_error(exc)
+            logger.error(
+                "{} LLM failure kind={} type={} user_id={} message={}",
+                _LOG_PREFIX,
+                error_kind,
+                type(exc).__name__,
+                user_id,
+                str(exc)[:500],
+            )
+            raise TopicRecommendationLLMError(
+                "Unable to generate topic recommendations from LLM",
+                error_kind=error_kind,
+            ) from exc
+
+    raw = _call_llm(user_prompt)
     logger.info(
         "{} Recommendations received user_id={} keys={}",
         _LOG_PREFIX,
@@ -372,12 +484,7 @@ def _call_llm_with_validation_retry(
 
     retry_prompt = f"{user_prompt}{VALIDATION_RETRY_USER_SUFFIX}"
     logger.info("{} Sending validation retry to LLM user_id={}", _LOG_PREFIX, user_id)
-    retry_raw = call_topic_recommendation_llm(
-        system_prompt=TOPIC_RECOMMENDATION_SYSTEM_PROMPT,
-        user_prompt=retry_prompt,
-        user_id=user_id,
-        generate_fn=generate_fn,
-    )
+    retry_raw = _call_llm(retry_prompt)
     logger.info(
         "{} Retry recommendations received user_id={} keys={}",
         _LOG_PREFIX,
